pre firebase integration
